Refactor Server to Rest, separate the routes
