+++ rest/src/lib.rs
@@ -17,36 +17,23 @@
 #[macro_use]
 extern crate tracing;
 
-pub mod error;
-pub use error::*;
+mod helpers;
+pub use helpers::*;
 
-pub mod server;
-pub use server::*;
+mod routes;
+pub use routes::*;
 
-use snarkvm_compiler::Transaction;
-use snarkvm_console::prelude::Network;
+use crate::{with, OrReject, RestError};
+
+use snarkvm_compiler::{BlockStorage, Ledger, ProgramStorage, RecordsFilter, Transaction};
+use snarkvm_console::{account::ViewKey, prelude::Network, types::Field};
 
 use anyhow::Result;
-use tokio::sync::mpsc;
-use warp::{reject, Filter, Rejection};
-
-/// A trait to unwrap a `Result` or `Reject`.
-pub trait OrReject<T> {
-    /// Returns the result if it is successful, otherwise returns a rejection.
-    fn or_reject(self) -> Result<T, Rejection>;
-}
-
-impl<T> OrReject<T> for anyhow::Result<T> {
-    /// Returns the result if it is successful, otherwise returns a rejection.
-    fn or_reject(self) -> Result<T, Rejection> {
-        self.map_err(|e| reject::custom(ServerError::Request(e.to_string())))
-    }
-}
-
-/// A middleware to include the given item in the handler.
-pub(crate) fn with<T: Clone + Send>(item: T) -> impl Filter<Extract = (T,), Error = std::convert::Infallible> + Clone {
-    warp::any().map(move || item.clone())
-}
+use indexmap::IndexMap;
+use parking_lot::RwLock;
+use std::sync::Arc;
+use tokio::{sync::mpsc, task::JoinHandle};
+use warp::{http::StatusCode, reject, reply, Filter, Rejection, Reply};
 
 /// Shorthand for the parent half of the `Ledger` message channel.
 pub type LedgerSender<N> = mpsc::Sender<LedgerRequest<N>>;
@@ -57,4 +44,88 @@ pub type LedgerReceiver<N> = mpsc::Receiver<LedgerRequest<N>>;
 #[derive(Debug)]
 pub enum LedgerRequest<N: Network> {
     TransactionBroadcast(Transaction<N>),
+}
+
+/// A REST API server for the ledger.
+pub struct Rest<N: Network, B: BlockStorage<N>, P: ProgramStorage<N>> {
+    /// The ledger.
+    ledger: Arc<RwLock<Ledger<N, B, P>>>,
+    /// The ledger sender.
+    ledger_sender: LedgerSender<N>,
+    /// The server handles.
+    handles: Vec<JoinHandle<()>>,
+}
+
+impl<N: Network, B: 'static + BlockStorage<N>, P: 'static + ProgramStorage<N>> Rest<N, B, P> {
+    /// Initializes a new instance of the server.
+    pub fn start(
+        ledger: Arc<RwLock<Ledger<N, B, P>>>,
+        additional_routes: Option<impl Filter<Extract = impl Reply, Error = Rejection> + Clone + Sync + Send + 'static>,
+    ) -> Result<(Self, LedgerReceiver<N>)> {
+        // Initialize a channel to send requests to the ledger.
+        let (ledger_sender, ledger_receiver) = mpsc::channel(64);
+
+        // Initialize a vector for the server handles.
+        let mut handles = Vec::new();
+
+        // Initialize the routes.
+        let routes = Self::routes(ledger.clone(), ledger_sender.clone());
+
+        // Spawn the server.
+        handles.push(tokio::spawn(async move {
+            let addr = ([0, 0, 0, 0], 80);
+
+            // Start the server with optional additional routes.
+            match additional_routes {
+                Some(additional_routes) => {
+                    warp::serve(routes.or(additional_routes)).run(addr).await;
+                }
+                None => {
+                    warp::serve(routes).run(addr).await;
+                }
+            }
+        }));
+
+        let server = Self { ledger, ledger_sender, handles };
+
+        Ok((server, ledger_receiver))
+    }
+
+    /// Initializes a ledger handler.
+    pub fn start_handler(
+        &mut self,
+        ledger: Arc<RwLock<Ledger<N, B, P>>>,
+        mut ledger_receiver: LedgerReceiver<N>,
+    ) -> JoinHandle<()> {
+        tokio::spawn(async move {
+            while let Some(request) = ledger_receiver.recv().await {
+                match request {
+                    LedgerRequest::TransactionBroadcast(transaction) => {
+                        let transaction_id = transaction.id();
+                        match ledger.write().add_to_memory_pool(transaction) {
+                            Ok(()) => trace!("✉️ Added transaction '{transaction_id}' to the memory pool"),
+                            Err(error) => {
+                                warn!("⚠️ Failed to add transaction '{transaction_id}' to the memory pool: {error}")
+                            }
+                        }
+                    }
+                };
+            }
+        })
+    }
+
+    /// Returns the ledger.
+    pub fn ledger(&self) -> Arc<RwLock<Ledger<N, B, P>>> {
+        self.ledger.clone()
+    }
+
+    /// Returns the ledger sender.
+    pub fn ledger_sender(&self) -> &LedgerSender<N> {
+        &self.ledger_sender
+    }
+
+    /// Returns the handles.
+    pub fn handles(&self) -> &Vec<JoinHandle<()>> {
+        &self.handles
+    }
 }
